Create levelinformationscript
This script initializes a scene upon loading. It holds data that is used by, but varies, from scene to scene; things like the displayed name, description/instructions, and the number of points required to advance.
Part of this code reads from and edits the PlayerPrefs file to check and update high scores for individual levels. The way scenes are divided and ordered in Gravity Gerbil, they need to be called by name rather than by number. Lines 42 and 43 look for, and create if necessary, a particular variable in PlayerPrefs containing the level name. When the program is updated and new levels are added or reordered, this section of code will automatically account for that with no further action on the developer's part.
Line 46 is an example of PlayerPrefs using multiple variables for a specific level.